+++ spec/components/image_card_spec.rb
@@ -123,4 +123,14 @@ describe "ImageCard", type: :view do
     render_component(href: "#", lang: "cy")
     assert_select ".gem-c-image-card[lang='cy']"
   end
+
+  it "applies lazy loading attribute when lazy is specified" do
+    render_component(href: "#", image_src: "/moo.jpg", image_alt: "some meaningful alt text", image_loading: "lazy")
+    assert_select ".gem-c-image-card__image[loading='lazy']"
+  end
+
+  it "checks image loading attribute is 'auto' when value 'image_loading' is not specified" do
+    render_component(href: "#", image_src: "/moo.jpg", image_alt: "some meaningful alt text")
+    assert_select ".gem-c-image-card__image[loading='auto']"
+  end
 end