+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 # Unreleased
 
+* Add lazy loading to image card ([PR #2270](https://github.com/alphagov/govuk_publishing_components/pull/2270))
 * GOVUK.Modules.start can accept DOM elements ([PR #2260](https://github.com/alphagov/govuk_publishing_components/pull/2260))
 
 # 25.3.1